// dashboard/src/app/core/api/api-client.ts
export enum SlotGenerationMode {
    Grid = 0,
    FixedStartTimes = 1,
}
// ...
export interface EmployeeScheduleDto {
    activeFrom?: Date;
    activeTo?: Date;
    numberOfCycles?: number;
    days?: EmployeeScheduleDayDto[];
    id?: string | undefined;
    slotGenerationMode?: SlotGenerationMode;
    isActive?: boolean;
}

export interface EmployeeScheduleDayDto {
    cycleIndex?: number;
    workRanges?: TimeRangeDto[];
    breaks?: TimeRangeDto[];
    fixedStartTimes?: string[] | undefined;
}

export interface TimeRangeDto {
    startTime?: string;
    endTime?: string;
}